backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0

# Async file operations
aiofiles>=23.2.1

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# File type detection
python-magic>=0.4.27

# PDF generation
weasyprint>=60.0

# Security (rate limiting)
slowapi>=0.1.9
